+++ package.json
@@ -17,6 +17,7 @@
     "date-fns": "^2.0.0-beta.5",
     "i": "^0.3.6",
     "material-ui-image": "^3.3.2",
+    "node-sass": "^6.0.1",
     "npm": "^7.19.0",
     "react": "^17.0.2",
     "react-bootstrap-validation": "^0.1.11",
@@ -24,9 +25,12 @@
     "react-device-detect": "^1.17.0",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.3.6",
+    "react-id-swiper": "^4.0.0",
+    "react-material-ui-carousel": "^2.3.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-spinners": "^0.11.0",
+    "swiper": "^6.8.4",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
